themes directory restructure
