+++ lib/src/logger.paw.dart
@@ -1,5 +1,5 @@
 import '../paw.dart';
-import 'themes/core/interface.theme.paw.dart';
+import 'themes/interface.theme.paw.dart';
 import 'utils/log_levels.utils.paw.dart';
 
 ///
