+++ lib/src/utils/utils.paw.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import '../colors/ansi.paw.dart';
-import '../themes/core/interface.theme.paw.dart';
+import '../themes/interface.theme.paw.dart';
 import 'log_levels.utils.paw.dart';
 
 ///
